move classes into Dto-domain packages
