+++ src/main/java/com/jobfinder/finder/dto/admin/AdminResponseDto.java
@@ -1,4 +1,4 @@
-package com.jobfinder.finder.dto;
+package com.jobfinder.finder.dto.admin;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,8 +9,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @Getter
 @Setter
-public class AddressDto {
-  private String country;
-  private String city;
-  private String postalCode;
+public class AdminResponseDto {
+  private String username;
+  private String email;
 }
