+++ src/main/java/com/jobfinder/finder/mapper/AdminMapper.java
@@ -1,7 +1,7 @@
 package com.jobfinder.finder.mapper;
 
-import com.jobfinder.finder.dto.AdminCreationDto;
-import com.jobfinder.finder.dto.AdminResponseDto;
+import com.jobfinder.finder.dto.admin.AdminCreationDto;
+import com.jobfinder.finder.dto.admin.AdminResponseDto;
 import com.jobfinder.finder.entity.AdminEntity;
 import org.mapstruct.Mapper;
