+++ src/main/java/com/jobfinder/finder/repository/PostRepository.java
@@ -1,14 +1,8 @@
 package com.jobfinder.finder.repository;
 
-import com.jobfinder.finder.dto.PostFilterRequestDto;
 import com.jobfinder.finder.entity.PostEntity;
-import java.awt.print.Pageable;
-import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.stereotype.Repository;
 
 public interface PostRepository extends JpaRepository<PostEntity, Long>, JpaSpecificationExecutor<PostEntity> {
 }
