+++ src/test/java/com/jobfinder/finder/integrationTest/controller/PostControllerIntegrationTest.java
@@ -4,11 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jobfinder.finder.constant.EmploymentType;
 import com.jobfinder.finder.constant.PostStatus;
 import com.jobfinder.finder.controller.PostController;
-import com.jobfinder.finder.dto.PostDto;
-import com.jobfinder.finder.dto.PostFilterRequestDto;
-import com.jobfinder.finder.entity.PostEntity;
+import com.jobfinder.finder.dto.post.PostDto;
+import com.jobfinder.finder.dto.post.PostFilterRequestDto;
 import com.jobfinder.finder.integrationTest.configuration.MockUserDetailsManagerConfig;
-import com.jobfinder.finder.repository.PostRepository;
 import com.jobfinder.finder.service.PostService;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -20,9 +18,6 @@ import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
